# docs/api/openapi.yaml
components:
  schemas:
    AccountBalanceOut:
      properties:
        account_id:
          title: Account Id
          type: string
        current_balance_cents:
          title: Current Balance Cents
          type: integer
        projected_balance_cents:
          title: Projected Balance Cents
          type: integer
      required:
      - account_id
      - current_balance_cents
      - projected_balance_cents
      title: AccountBalanceOut
      type: object
    AccountCreate:
      additionalProperties: false
      properties:
        currency:
          maxLength: 3
          minLength: 3
          title: Currency
          type: string
        initial_balance_cents:
          default: 0
          minimum: 0.0
          title: Initial Balance Cents
          type: integer
        kind:
          title: Kind
          type: string
        name:
          maxLength: 255
          minLength: 1
          title: Name
          type: string
      required:
      - name
      - kind
      - currency
      title: AccountCreate
      type: object
    AccountOut:
      properties:
        archived_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Archived At
        created_at:
          format: date-time
          title: Created At
          type: string
        currency:
          title: Currency
          type: string
        current_balance_cents:
          title: Current Balance Cents
          type: integer
        current_balance_version:
          title: Current Balance Version
          type: integer
        id:
          format: uuid
          title: Id
          type: string
        initial_balance_cents:
          title: Initial Balance Cents
          type: integer
        kind:
          title: Kind
          type: string
        name:
          title: Name
          type: string
      required:
      - id
      - name
      - kind
      - currency
      - initial_balance_cents
      - current_balance_cents
      - current_balance_version
      - archived_at
      - created_at
      title: AccountOut
      type: object
    AccountUpdate:
      additionalProperties: false
      properties:
        archived:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Archived
        expected_version:
          anyOf:
          - minimum: 0.0
            type: integer
          - type: 'null'
          title: Expected Version
        kind:
          anyOf:
          - type: string
          - type: 'null'
          title: Kind
        name:
          anyOf:
          - maxLength: 255
            minLength: 1
            type: string
          - type: 'null'
          title: Name
      title: AccountUpdate
      type: object
    AttachmentOut:
      properties:
        account_id:
          format: uuid
          title: Account Id
          type: string
        content_type:
          title: Content Type
          type: string
        created_at:
          format: date-time
          title: Created At
          type: string
        file_name:
          title: File Name
          type: string
        id:
          format: uuid
          title: Id
          type: string
        size_bytes:
          title: Size Bytes
          type: integer
      required:
      - id
      - account_id
      - file_name
      - content_type
      - size_bytes
      - created_at
      title: AttachmentOut
      type: object
    Body_import_account_csv_accounts__account_id__imports_post:
      properties:
        file:
          contentMediaType: application/octet-stream
          title: File
          type: string
        source_name:
          title: Source Name
          type: string
      required:
      - file
      - source_name
      title: Body_import_account_csv_accounts__account_id__imports_post
      type: object
...
    Body_upload_attachment_accounts__account_id__attachments_post:
      properties:
        file:
          contentMediaType: application/octet-stream
          title: File
          type: string
      required:
      - file
      title: Body_upload_attachment_accounts__account_id__attachments_post
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ImportBatchOut:
      properties:
        account_id:
          format: uuid
          title: Account Id
          type: string
        created_at:
          format: date-time
          title: Created At
          type: string
        file_name:
          title: File Name
          type: string
        file_sha256:
          title: File Sha256
          type: string
        id:
          format: uuid
          title: Id
          type: string
        imported_count:
          title: Imported Count
          type: integer
        row_count:
          title: Row Count
          type: integer
        skipped_count:
          title: Skipped Count
          type: integer
        source_name:
          title: Source Name
          type: string
      required:
      - id
      - account_id
      - source_name
      - file_name
      - file_sha256
      - row_count
      - imported_count
      - skipped_count
      - created_at
      title: ImportBatchOut
      type: object
...
    SnapshotCreate:
      additionalProperties: false
      properties:
        note:
          anyOf:
          - maxLength: 500
            minLength: 1
            type: string
          - type: 'null'
          title: Note
        reported_balance_cents:
          title: Reported Balance Cents
          type: integer
      required:
      - reported_balance_cents
      title: SnapshotCreate
      type: object
    SnapshotOut:
      properties:
        account_id:
          format: uuid
          title: Account Id
          type: string
        created_at:
          format: date-time
          title: Created At
          type: string
        difference_cents:
          title: Difference Cents
          type: integer
        id:
          format: uuid
          title: Id
          type: string
        ledger_balance_cents:
          title: Ledger Balance Cents
          type: integer
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        reported_balance_cents:
          title: Reported Balance Cents
          type: integer
        status:
          title: Status
          type: string
      required:
      - id
      - account_id
      - reported_balance_cents
      - ledger_balance_cents
      - difference_cents
      - status
      - note
      - created_at
      title: SnapshotOut
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: auth/login
      type: oauth2
info:
  title: Cifra API
  version: 0.2.0
openapi: 3.1.0
paths:
  /accounts:
    get:
      operationId: list_accounts_accounts_get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AccountOut'
                title: Response List Accounts Accounts Get
                type: array
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: List Accounts
      tags:
      - accounts
    post:
      operationId: create_account_accounts_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Create Account
      tags:
      - accounts
  /accounts/{account_id}:
    delete:
      operationId: delete_account_accounts__account_id__delete
      parameters:
      - in: path
        name: account_id
        required: true
        schema:
          format: uuid
          title: Account Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Delete Account
      tags:
      - accounts
    get:
      operationId: get_account_accounts__account_id__get
      parameters:
      - in: path
        name: account_id
        required: true
        schema:
          format: uuid
          title: Account Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Get Account
      tags:
      - accounts
    patch:
      operationId: update_account_accounts__account_id__patch
      parameters:
      - in: path
        name: account_id
        required: true
        schema:
          format: uuid
          title: Account Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Update Account
      tags:
      - accounts
  /accounts/{account_id}/attachments:
    post:
      operationId: upload_attachment_accounts__account_id__attachments_post
      parameters:
      - in: path
        name: account_id
        required: true
        schema:
          format: uuid
          title: Account Id
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_attachment_accounts__account_id__attachments_post'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Upload Attachment
      tags:
      - accounts
  /accounts/{account_id}/attachments/{attachment_id}:
    get:
      operationId: download_attachment_accounts__account_id__attachments__attachment_id__get
      parameters:
      - in: path
        name: account_id
        required: true
        schema:
          format: uuid
          title: Account Id
          type: string
      - in: path
        name: attachment_id
        required: true
        schema:
          format: uuid
          title: Attachment Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Download Attachment
      tags:
      - accounts
  /accounts/{account_id}/balance:
    get:
      operationId: account_balance_accounts__account_id__balance_get
      parameters:
      - in: path
        name: account_id
        required: true
        schema:
          format: uuid
          title: Account Id
          type: string
      - in: query
        name: projected
        required: false
        schema:
          default: false
          title: Projected
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountBalanceOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Account Balance
      tags:
      - accounts
  /accounts/{account_id}/imports:
    post:
      operationId: import_account_csv_accounts__account_id__imports_post
      parameters:
      - in: path
        name: account_id
        required: true
        schema:
          format: uuid
          title: Account Id
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_import_account_csv_accounts__account_id__imports_post'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportBatchOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Import Account Csv
      tags:
      - accounts
  /accounts/{account_id}/snapshots:
    get:
      operationId: list_snapshots_accounts__account_id__snapshots_get
      parameters:
      - in: path
        name: account_id
        required: true
        schema:
          format: uuid
          title: Account Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SnapshotOut'
                title: Response List Snapshots Accounts  Account Id  Snapshots Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: List Snapshots
      tags:
      - accounts
    post:
      operationId: create_snapshot_accounts__account_id__snapshots_post
      parameters:
      - in: path
        name: account_id
        required: true
        schema:
          format: uuid
          title: Account Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SnapshotCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Create Snapshot
      tags:
      - accounts
...